Implemented code review comments
