+++ java/src/main/java/com/codurance/training/tasks/TaskList.java
@@ -4,26 +4,22 @@ import com.codurance.training.tasks.handler.TaskExecutionHandlerImpl;
 
 import java.io.BufferedReader;
 import java.io.IOException;
-import java.io.InputStreamReader;
 import java.io.PrintWriter;
-import java.util.LinkedHashMap;
-import java.util.List;
-import java.util.Map;
 
 public final class TaskList implements Runnable {
     private static final String QUIT = "quit";
+    private BufferedReader in ;
+    private final PrintWriter out;
+    private final TaskExecutionHandlerImpl taskExecutionHandler;
 
-    private final TaskExecutionHandlerImpl taskExecutionHandler = new TaskExecutionHandlerImpl();
 
-
-    public TaskList() {
-        //default constructor without parameter
+    public TaskList(BufferedReader in, PrintWriter out) {
+        this.in = in;
+        this.out = out;
+        this.taskExecutionHandler = new TaskExecutionHandlerImpl(out);
     }
 
     public void run() {
-        BufferedReader in = new BufferedReader(new InputStreamReader(System.in));
-        PrintWriter out = new PrintWriter(System.out);
-
         while (true) {
             out.print("> ");
             out.flush();
