+++ java/src/main/java/com/codurance/training/tasks/handler/TaskExecutionHandlerImpl.java
@@ -1,30 +1,52 @@
 package com.codurance.training.tasks.handler;
 
-import com.codurance.training.tasks.actions.TaskActionHandlerImpl;
+import com.codurance.training.tasks.model.Task;
+import com.codurance.training.tasks.service.*;
+
+import java.io.PrintWriter;
+import java.util.LinkedHashMap;
+import java.util.List;
+import java.util.Map;
 
 public class TaskExecutionHandlerImpl implements TaskExecutionHandler{
-    private final TaskActionHandlerImpl taskActionHandlerImpl = new TaskActionHandlerImpl();
+    private static Map<String, List<Task>> tasks;
+
+    private static long lastId;
+    private SelectionService selectionService;
+
+    private DisplayService displayService;
+
+    private AddService addService;
+    public TaskExecutionHandlerImpl(PrintWriter out){
+        tasks = new LinkedHashMap<>();
+        lastId = 0;
+        this.selectionService = new SelectionServiceImpl(out, tasks);
+        this.displayService = new DisplayServiceImpl(out, tasks);
+        this.addService = new AddServiceImpl(out, tasks, lastId);
+    }
+
     public void execute(String commandLine){
+
         String[] commandRest = commandLine.split(" ", 2);
         String command = commandRest[0];
         switch (command) {
             case "show":
-                taskActionHandlerImpl.show();
+                this.displayService.show();
                 break;
             case "add":
-                taskActionHandlerImpl.add(commandRest[1]);
+                this.addService.add(commandRest[1]);
                 break;
             case "check":
-                taskActionHandlerImpl.check(commandRest[1]);
+                this.selectionService.check(commandRest[1]);
                 break;
             case "uncheck":
-                taskActionHandlerImpl.uncheck(commandRest[1]);
+                this.selectionService.uncheck(commandRest[1]);
                 break;
             case "help":
-                taskActionHandlerImpl.help();
+                this.displayService.help();
                 break;
             default:
-                taskActionHandlerImpl.error(command);
+                this.displayService.error(command);
                 break;
         }
     }
