+++ java/src/main/java/com/codurance/training/tasks/model/Task.java
@@ -1,4 +1,4 @@
-package com.codurance.training.tasks;
+package com.codurance.training.tasks.model;
 
 public final class Task {
     private final long id;
