+++ java/src/main/java/com/codurance/training/tasks/service/AddService.java
@@ -1,0 +1,7 @@
+package com.codurance.training.tasks.service;
+
+public interface AddService {
+
+    void add(String commandLine);
+
+}
