+++ java/src/main/java/com/codurance/training/tasks/service/DisplayService.java
@@ -1,0 +1,9 @@
+package com.codurance.training.tasks.service;
+
+public interface DisplayService {
+    void help();
+
+    void show();
+
+    void error(String command);
+}
