+++ java/src/main/java/com/codurance/training/tasks/service/SelectionService.java
@@ -1,0 +1,10 @@
+package com.codurance.training.tasks.service;
+
+public interface SelectionService {
+
+    void check(String idString);
+
+    void uncheck(String idString);
+
+    void setDone(String idString, boolean done);
+}
